Source typo fix (includes file rename)
